requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
lxml==6.0.2
packaging==26.0
pillow==12.1.1
typing_extensions==4.15.0
xlsxwriter==3.2.9
